+++ examples/storybook/stories/Wrapper/Themes.js
@@ -1,13 +1,12 @@
 import React from 'react';
-import { RichContentEditorBox, Page, Section, ContentState } from '../Components/StoryParts';
+import { Page, Section, ContentState } from '../Components/StoryParts';
 import exapmleState from '../../../../e2e/tests/fixtures/themeing-info.json';
 import Palette from '../Components/Palette';
 import { wixPalettes } from '../palettesExample';
 import SegmentedToggle from 'wix-style-react/SegmentedToggle';
 import FormField from 'wix-style-react/FormField';
 import { Layout } from 'wix-style-react/Layout';
-import EditorWrapper from '../Components/EditorWrapper';
-import sourcecode from '!!raw-loader!../Components/EditorWrapper';
+import ExampleApplication from '../Components/ExampleApplication';
 
 export default () => {
   return (
@@ -50,16 +49,12 @@ class ThemeSelector extends React.Component {
             </SegmentedToggle>
           </FormField>
         </Layout>
-
         <Palette palette={wixPalettes[selected]} />
-
-        <RichContentEditorBox sourcecode={sourcecode}>
-          <EditorWrapper
-            key={selected}
-            contentState={exapmleState}
-            palette={wixPalettes[selected]}
-          />
-        </RichContentEditorBox>
+        <ExampleApplication
+          key={selected}
+          initialState={exapmleState}
+          palette={wixPalettes[selected]}
+        />
       </React.Fragment>
     );
   }